[CDS-118] Typography use css variables (#6535)
* chore(codegen): rename codegen/typography -> codegen/main

* build: augment base.package.json

* refactor(Text): use global CSS variables and remove ScaleProvider

* make scale classes and CSS vars not global

We want to support having more than one scale on a page in v1 to
make product onboarding to CDS easier. The primary use case is
for products that use multiple scales like normal and dense on
different pages, the nav bar should only be at one scale. This results
in some pages having to have more than one scale.

* rename typographies to typography